chore: update the use of enum
Move the method that convert the bridge number to bridge position from BridgeMaker to BridgeInformation. This function should position where the data exist. Implementation and name of the method is also changed {
Implementation: Originally it was void that update the instance variable bridgePosition. However, now it return the BridgeInformation object according the bridgeNumber.
name: Because it's implementation is changed, renamed it to convertToBridgeInfo.
}
Implement getPosition method to return bridgePosition on Enum.