refactor(BridgeGame): modify printResultService() method parameters
Before change the printResultService method generate game screen data
and game success or failure value with the input parameters and handing it over to the controller

After change that method receive the game screen data and game success
or failure value directly. As a result, it is possible to increase the readability of the code.